Add support for the scanning API (#74)
* Refactor secure and monitor into it's own files
_client.py has grown too big. Moving SdcMonitor and SdcSecure into
separated files.

* Fix python style (pep8)
Make the code pep8 compilant except for line lenght (E501) and for
comparisons to none (E711) and bool (E712) using '==' or '!='.

* Add SdScanningClient class with the secure scanning API
Add image, registries and subscription support. Importing some anchore
helper functions from anchore-cli into sdcclient/_anchore.

* Add support for scanning policies

* Add support for scanning alerts and runtime
